requirements.txt
# AI Core Dependencies
pydantic>=2.6.3
pydantic-settings>=2.1.0
zhipuai>=2.1.5
langchain>=0.0.350
langchain-community>=0.0.10
langchain-core>=0.1.4
python-dotenv>=1.0.0

# FastAPI Dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
python-jose>=3.3.0
passlib>=1.7.4
httpx>=0.27.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
aiosqlite>=0.19.0

# Utils
loguru>=0.7.2
pytest>=7.4.3
openpyxl>=3.1.2

# AI
pillow==10.1.0
